refactor(package): app 패키지 구성
