+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.app.LottoApp;
+
 public class Application {
     public static void main(String[] args) {
         LottoApp lottoApp = LottoAppDownloader.downLoad();
